feat: mandatory nr averaged measurements
For a MeasureControllerABC, make the number of measurements to average over mandatory. This is necessary as we're setting this value to 1 externally in continuous time-resolved measurements (where we are not supposed to know about implementation details of controllers).

Also, give MeasureControllerABC a property getter, and remove it from the viperino implementation.