fix: prevent running multiple bad_px_finders
Make sure only one bad pixel finder can run at the same time. This includes the case in which a finder is not yet deleted (or garbage-collected) when the user clicks "find" again.

Also:
(1) decorate with pyqtSlot one more method that can execute in the wrong thread: __on_progess, connected to __finder.report_progress, since __finder is moved to a different thread
(2) Abort when an error occurs