fix: bad pixel finder - multiple bugs
Multiple bugs appeared after methods were decorated with @pyqtSlot in commit 8469617cfacae475fcf49ba490d8d1ee56eb4723

Before that, it is now quite clear that the slots were running in the wrong thread (likely the GUI thread) even if the finder object is moved to a different thread.

The correctness of the execution thread, however, caused also methods of GUI objects (self.__info) or objects from other threads having timer children (__camera) to be executed now in the finder thread rather than in their own threads. To get stuff right, use signals to .start, .trigger, and .set_settings of the camera, as well as .setInformativeText and .exec_ the info dialog.

Also, get the info from the camera at instantiation, so it happens only once; and take care of correctly reporting progress (in case changes of settings are not yet applied).

Finally, since now the finder is correctly running in its own thread, there is no need to have an extra QThread run calculations.