feat: new class Version
Simple class that handles version comparisons. Accepts (major, minor, patch) with minor and patch optional, and strings of the form "<major>[<minor>[.<patch>]]" with "[]" optional. Supports equality and ordering comparisons (>, <, >=, <=, ==, !=), also with respect to a string of the same form as accepted at instantiation.

Attributes: .major, .minor, .patch. The former is always present. Querying the others may raise RuntimeError if not originally known.