feat: thermocouple temperature °C
ViPErInoController now correctly converts the thermocouple temperature to °C, if an appropriate 'conversions'/'thermocouple_type' is given, and if it is also possible to measure the LM35 cold-junction temperature.

Still to be done: cross check that the temperature values are actually CORRECT. Probably good to test with LN2, which should still be in range.