fix: prevent AttributeError on invalid settings
When creating a measurement with invalid settings, we may end up having self.serial == None for some controllers. Trying to access any attribute of this raises AttributeErrors rather than finishing gracefully.

In this case, it can also happen that settings is empty, so it could have no sections/options to directly access. Attempting this in the finalization of TimeResolved continuous (calling set_continuous_mode) would crash with NoSectionError. The problem is, of course, that finalization happens also in case of errors.