fix: improve energy calibration fit
Add a few checks on the computed energy calibration:
- require at least 10 measurements to perform the fit
- make sure offset and gain coefficients have reasonable values, not too far from (0, 1). Accept offsets within +-50(eV) and gains between 1e-3 and 100. These ranges are VERY CONSERVATIVE, and can possibly narrowed a bit more at later times.
- Check that the sum of residuals is reasonably small. This is also an indicator of a reasonable fit.

Eventually we will display the fit and the residuals plot to the user, who will choose whether to save the calibration curve. The limits above are expected to be violated if the user, e.g., forgot to plug the cable (or if something is wrong with either the LEED optics or the measuring hardware).