fix: abort did not set energy to zero
When aborting a measurement at the end of the first step of preparation, i.e., while waiting for the "OK" returning after the first "set_voltage", the .__force_end timer could time out first, terminating the measurement without proper cleanup (including setting energy to zero).

The solution is not waiting much for a missing OK when a controller should stop. Right now we wait at most 200 ms (which should be enough for sending several quick commands).

This allows stopping also in the middle of a calibration.

It should still be tested for the case of multiple controllers.