feat: Class for Delta file
One single delta file contains a large amount of information that can not be stored in a single array. This class takes care of reading in the file and holding all the information in a more suitable object. Note that ever Delta file is associated with a single Atom, but an Atom can have more than one Delta file.

Co-Authored-By: Michele Riva <riva@iap.tuwien.ac.at>