fix: ISCamera on non-windows errors
Raise understandable errors if someone tries to import an imagingsource camera (or launch the measurement part) from a non-windows environment.

An even better solution would be to deactivate camera related stuff, but let users use the rest.

This will be necessary when I have the pattern simulator back up and running.