fix: AUXEXPBEAMS  scaling error
In writeAUXEXPBEAMS the (pseudo-)experimental intensity values are written with a normalization to the file AUXEXPBEAMS. This normalization was wrongly applied directly to the beam object rather than just the output. Therefore if multiple R-factor calculations are run (e.g. in the error calc) this normalization would be carried through. This mostly affects very low intensity beams due to rounding errors.