feat: more flexible device config finder
Add optional argument tolerant_match to hardwarebase.get_device_config. This allows to switch between exact match (False, old behavior) and 'match apart from stuff between square brackets'. This has the huge advantage that we can now realize that an ImagingSourceCamera is unchanged even if there is a user-defined name (on top of the model/serial/etc).

This is now the default. However, we should consider the corner cases in which the device name in the configuration file saved to disk does not match the unique name id of the camera (stupidly the implementation of the IS driver adds the user-defined name if given).

ViPErLEEDSettings.from_settings will however still behave the old way by default.