fix: update device_name when selected
Update configuration file for cameras the same way as we do for controllers when a user picks from the device list. This helps limiting a bit the "device not found" errors when a measurement is started. Eventually I want to correct these each time a measurement is started based on the known devices.