fix: change R=0 error to warning
Previously an error would be raised by tleedm if a R == 0 is detected because this usually points to an error in the calculation.
However, as pointed out by Lutz, when comparing with pseudo-experiment data, a R-faktor of zero may be valid.

This changes the error to a warning and does not stop execution anymore.