fix: last-resort close all widgets
When the selector plugin is closed, make sure all top-level widgets of the application are also closed. This is in case any module fails to successfully close all its windows.

This should make sure that the QApplication terminates when the selector is closed.

This temporarily solves issues with CameraViewer, which, under some non-reproducible circumstances can still pop up out of nowhere.