fix: CameraViewer on BadPx
The previous implementation of BadPixelsFinder was creating a CameraViewer in the hope it could close one that may have been open already pointing to the same camera. However, this solution is entirely useless, as the CameraABC instances that the finder gets are NOT THE SAME as those in the UI. Hence the trick did not work at all. To make it work, one would need to have "unique" CameraABC instances, perhaps name based. This would generally be good to solve the issue of accessing the same hardware from two different instances.

Should we ever decide to go that route, we can undo some of the edits in this commit.

As a result of this miscoding, opening the BadPxFinder with an active camera (running and visible) crashed miserably as we were attempting access to the same device.

The brute-force solution here simply forces all known cameras to stop before opening the finder.