fix: badpx, flat: oscillations in exposure time
Try handling the case of oscillatory intensity (and, hence, exposure times) between successive frames acquired in 'flat' mode.

This happened to me while helping installation in Julich (Max Buchta, Wuttig group).

The current solution keeps track of the 'history' of adjustments, detects when there is an 'essentially settled, just oscillating' situation, and returns an average 'exposure' to be used. The detection of 'settled' is done using a Holt-Winters filter on the 5-point median-smoothed data. The average is calculated on all raw datapoints after the last 'settling', and returned if this is reasonably small compared to the standard error of the mean.