feat: refactors and attempts to fix bug
We encounter a bug with very noisy intensities for high LMAX and low energies that seems to be inherent to TensErLEED.
The bug only seems to appear for composite layers.
Tried to replace subroutine CXMTXT in TAUMAT with BLAS calls, but work in progress...
Currently LU factorization returns detects singular matrix.

Co-Authored-By: Michele Riva <riva@iap.tuwien.ac.at>