feat: imagingsource black_level
Allow accessing the "Brightness" or "Black Level" property of GigE cameras.

The current implementation does not take into account that some ImagingSourceCameras may not have the black level property. Currently all GigE cameras seem to support it, though.

It would be relatively easy to extend to check whether the camera supports it.