feat: CalibrationTask
New class that abstract away part of the functionality currently implemented in BadPixelsFinder to be used in more general cases of device-calibration operations.

In addition to the abstract class for calibration tasks, there is also one for corresponding Operations (IntEnum subclass) that can be used, at least, for progress reporting.