feat: CameraABC.calibration_tasks
New attribute of CameraABC instances that can be used to specify lists of CameraCalibrationTask subclasses to be executed, in insertion order, right before a camera is started, and right before running a bad-pixels-finding routine.

The former list is currently unused, and will be probably implemented after I figure out how to move a ImagingSourceCamera safely to another thread [probably recreating driver after moveToThread]. This is likely needed because right now the camera starting procedure is already very slow.