feat: SettingsDialog.settings_saved
Added signal emitted every time a user has changed settings, applied some, closed the dialog (by OK or Cancel or X), and decided not save changes to file.

Users can connect to the signal and should normally reset the device to the settings it had before the dialog was opened.

Also in this commit: handler tries to infer from type only if given, and tries from value otherwise.  Then complains if it could not figure out a default handler.