feat: RegionOfInterest.roi_changed
Signal is emitted whenever the region of interest rubber-band rectangle is edited.

Notice the isVisible() checks: The widget may receive moveEvent (but probably not resizeEvent) even when it has not been shown. When not shown yet, sizes and positions may be wrong, as show() causes all pending moving/resizing event to actually take place.

Notice also that the signal is NOT emitted in setGeometry for two reasons: (i) positions and sizes are NOT YET RIGHT by when this method ends, as they are actually edited in the event loop; (ii) at least one of moveEvent or resizeEvent will be called soon after setGeometry.

The current implementation has still a double-emission issue when the geometry change involves both a movement and a resizing.