feat: MappedComboBox
A combo box that stores data, and can show text that is different from its data. Most importantly, it provides direct access to the data with set_current_data.

While showing text and keeping different data is possible with the default model/view, direct write access to the data is not possible without subclassing QComboBox.