feat, fix: default for SYMMETRY_FIND_ORI
Set the default for SYMMETRY_FIND_ORI to True, independent of whether the slab was preprocessed or not. Therefore, a full symmetry redetection will always run, but this is anyway cheap for any reasonable system.

There was also a bug that made it impossible to set SYMMETRY_FIND_ORI = True when the slab was recognized as preprocessed, but this is obsolete as of this change.