fix: thread-safety in previous commit
The previous implementation of the exception hook handler had severe issues when an exception would occur in a non-GUI thread.

In fact, it would attempt to access the QMessageBox from the non-main thread, causing a stalling and solid crash.

The new implementation makes sure that execution of GUI-related stuff (the message box) happens in the main thread, returning then control to the calling thread after the user acknowledges the issue.