fix, feat: CameraViewer flags
Generalize creation of flags in CameraViewer to clean up code a bit.

Also, very importantly, have CameraViewer.__init__ ALWAYS take care of optional falgs that may be passed at instantiation. In fact, skipping flags would not allow to change the behavior on an cached instance by "creating it" (i.e., pulling it from the cache) with CameraViewer(**new_flags). This is important as, when a measurement is started, we do want to disable certain features (like stopping the camera if the viewer is closed).