fix: bug in previous commit
Turns out that getattr(obj, property) returns the property value, as one would expect for a normal attribute, and not the property object as someone said on StackOverflow. this means one should not sue fget/fset, but rather getattr/setattr directly. Also, setattr does the right thing of invoking the setter method.