fix: image underflow error
When calculating the saturation mask, computing img_array - _min CAN LEAD TO UNDERFLOW of the  uint16 type.

While this should not happen in principle, we have observed this problem in the 174e camera from Julich. This means that I may be calculating wrongly the minimum intensity returned in camera.intensity_limits. The 174e in Julich can return 0 intensities while being a 12bit camera using the higher 12 bits of 16. This is different from the 265 I have been testing on. This may have been a fix introduced at the firmware level by The Imaging Source recently.