fix: bad pixels - prevent UI stall
Obsessively clicking on the "abort" button while the camera is being set up would cause a very funny issue of UI stalling (probably trying to access the "info_box" from a non-UI thread). Most likely a case of queued vs. direct connection for event processing.

The solution is to not allow aborting the process during camera preparation, and turn the abort button "enabled" only when a task is just started, and disabled as soon as it is finished.