fix: CameraCalibration swallow errors
Unfortunately turns out that the fix in commit 9c02b52092e652f3496ad7a8af8d1a465412ca65 did not really work: the error_info passed is always a tuple because we format its message, thus it never has a .name attribute.

The solution is to move the method to CameraABC. This has the advantage of not requiring to import extra modules in cameracalibration.