feat: ImageViewer painting
Before this commit, ImageViewer would
(i) smoothly scale .pixmap() up. This has two effects: slowing down painting, and preventing one to look at individual pixels
(ii) paint the whole Pixmap even if only one region was visible. This is the default behavior of QLabel, as pixmaps are normally small, and used primarily for icons. This made the painting extremely small at large zoom factors.

The core of the problem in the default QLabel implementation was to produce a .scaled version of a QImage derived from .pixmap(), then convert it to another QPixmap (that is cached to speed up multiple repaints). The .scaled transformation was done with SmoothTransform.

For us: (i) we change the pixmap almost every single time we repaint [frames from camera]; (ii) we want to see pixels when zooming in; (iii) it does not make sense to cache images as they can become very large when zooming in considerably.

The reimplementation of paintEvent in this commit takes care of all these issues. Notice that we do not even paint any text in the QLabel. The trick is not to cache images (may slow down a tad painting for small frame rates and if the user moves/rescales/pans the image around very often) and rather paint only the area that is visible using event.rect().