fix: CameraViewer roi always invalid
Issue was introduced with the presence of the settings editor. In fact, this has constantly access to the current settings in the camera (rather than a copy). Thus camera.roi is not the roi that is currently applied in the camera while the ROI is being modified by the user.