fix: LMAX > 16 in PARAM file
Bug reported by Lutz:
Delta calculation would fail for LMAX>16 (if actually used; not just as upper range).
The possible values for the variable MNLMB in PARAM are hardcoded and were only defined up to maximum LMAX of 16.
Added additonal values up to LMAX=18. They are tabulated in delta.f