fix: re-add flag -fallow-argument-mismatch for mpifort
Newer versions of gfortran (and thus mpifort) will raise an error when compiling the search without the flag -fallow-argument-mismatch.
This is the case for gfortran packaged with GCC 10 and higher.

We removed the flag earlier because, unfortunately, older version of gfortran may not recognize the flag.
Faced with this issue however, it's better to support the newer compilers and re-add the flag by default.
We added a warning in the docs for FORTRAN_COMP concerning this.

Co-Authored-By: Michele Riva <riva@iap.tuwien.ac.at>
Co-Authored-By: Florian Kraushofer <66359721+fkraushofer@users.noreply.github.com>