fix, feat: uint underflow issues
The CameraCalibrationTask base class implementation remembers the "intensity limits" when its instances are __init__ed. However, for ImagingSourceCamera the minimum MAY BE INCORRECT.

The crucial fix was introduced in commit 8b7756e31a5758068b6c9e691e4442c95197574a, but this requires that one frame containing pixels with the minimum intensity has been acquired.

This happens AFTER the abovementioned classes are created. In turn, this potentially causes issues for the bad-pixel-detection, and dark-level-adjustment routines that often do <intensity> - <min_intensity> or <image> - <min_intensity>.

The smaller of the problems is subtracting a constant (16) rather than zero from floating-point values. The worst is, however, subtracting the same constant from uintXX values that are potentially zero --> make them shoot up to max(uintXX) - constant.

Fix: prepend an extra section to the dark-level task (that runs on the same camera object as the bad pixels following it) that acquires a single dark frame with very short exposure, gain, and "black level". These are the best conditions to actually get at least one "zero intensity" pixel.

The acquisition is skipped if we already know that the camera does return zero as its minimum pixel intensity.