feat: CameraViewer always-active actions
Allow some innocent actions to be changed also when all others are deactivated.

"Innocent" actions are those that do not alter camera properties but only image display.

One such action currently present is "auto-contrast", which is now not forced to anything when a measurement runs: users can still decide to activate it or not at runtime.