feat: Improve performance of symmetry finder algo
Fundamentally changes the algorithm used to detect plane symmetry. The old algorithm scales with O(n^2), this new version scales with O(n), where n is the number of atoms in the lowest-occupation sublayer.

In the old version, candidate points for rotations/mirrors/glides were generated from pairs and trios of atoms on a (2x2) supercell. Then, all candidate symmetry operations were tested in these positions. In the new algorithm, each symmetry operation is applied to the slab just once, then the slab is tested for 2D translation symmetry. The only candidates for translation symmetry are atom-to-atom translations, and it is sufficient to test translation vectors linking an arbitrary atom in the lowest-occupancy sublayer to every atom in that sublayer (including itself). The position of the rotation axis or mirror/glide plane can then be determined based on the translation vector.

Also fixed several small bugs with bulk symmetry detection:
- bulk doubling was not always performed when required
- bulk translation symmetry was compared "element-blind", so e.g. ordered alloys were assigned higher symmetry than they actually have.