fix: ROIEditor relative validity of values
Increasing the "x position" should reduce the maximum "width" available (and viceversa). This was not taken into consideration earlier, leading to clear problems. The fix now highlights the problematic values, but pressing "ok" or "apply" in the settings dialog is still possible. Added a note on how to fix this as well.