fix: IS driver catch null-pointer error
Looks like IC_GetUniqueNamefromList can error out returning a pointer to NULL also if index is in range. I could not find the corresponding cpp source code, so it is hard to tell where the problem comes from.

For now, print out the error text and some info.

Behavior happened during a test in Prague, where they have multiple (3) Imaging Source cameras, one of which is a USB camera [my guess: the problematic camera is the USB one, which was in use via Zoom while testing].