fix: name_clean for viperleedcontroller
The default of ControllerABC is to return a cleaned-up version self.name. Unfortunately, this does not work nicely for the viperleed controller as .name fetches the serial number from the .hardware information and falls back onto "UNKONWN_SERIAL_NR" which is not great in a file name. Also, the information may be present as part of the device name in the settings file. Rather use this information to produce a name_clean.

Perhaps a nicer solution would have been to always cross check that hardware info and device_name match