fix: Bad pixels not starting
Fixed a major bug that prevented the bad-pixels finding routines to start when a CameraViewer was opened beforehand.

The problem was that the camera that viewers handle had their error_occurred signal always connected to the viewer.close(), which, in turn, always ends up calling closeEvent() after the OS handles the event itself.

In turn, running the closeEvent() on a viewer that was set up with .stop_on_close = True would stop the camera at the first error, even if the viewer was hidden.

During bad pixels, we always set the bad-pixels path to an invalid value, as we need to run all the procedures on unprocessed images. This, however, produces error_occurred --> stop camera before anything can happen.

A better solution that the current one: when we will have error levels, close viewer only on fatal errors