fix: bug in plotting of R+var(R) in Errors.pdf
Tilman reported an issue in the Errors.pdf plotting: The level of R+var(R) may be different if multiple modes are calculated at the same time vs separately. This happens if the modes have different R_min (e.g. geometric and occ if not ran with best configuration with both).
Change write_errors_pdf() such that it doesn't take var as input, but rather calculates it new for each mode.