fix: avoid deadlock in excepthook
The previous implementation of the excepthook could result in a deadlock situation:
If a non-blocking asynchronous call in a non-GUI thread results in an exception, and it is followed by a blocking call (without returning to the GUI event loop in between), the second blocking call would wait forever while the non-GUI thread is locked because of the exception hook. However, the exception hook can never fully execute (and show the message box) because event processing never returns to the GUI thread.

The solution is remove the thread-specific connection introduced in d9ef4b502661623d6ac3b4058f365a9565520f5c.

Disadvantage: the non-GUI thread will keep running when the exception is raised (bull will soon after terminate, once the exception hook finishes executing).