fix: Camera settings when not connected
Disable the "Properties" action when a camera is not connected, and keep track of the possibility to enable it when a camera is connected back.

What still does not work: if a camera is lost, this information is not reflected on the enabled state of the action. Probably the only way would be to try and fail gracefully, telling the use what the problem is.