fix: slab transform apply only at left
The previous implementation of apply_matrix_transform would pre- and post-multiply ucell by matrix and inv(matrix). It would also apply matrix to the fractional positions.

This was obviously WRONG:
- the TUT^-1 effectively changes the unit cell vectors, unless T is a permutation matrix or a very simple rotation (only +-1 elements), and the unit cell is diagonal.
- Applying T to fractional positions makes no sense, as a "basis-change" transform should not touch the fractional coordinates (i.e., the positions of atoms within the unit cell).

The right way to do this is transforming Cartesian atomic coordinates. If c is the Cartesian coordinate of an atom (in the non-LEED reference frame) as a column vector, then we want to get c' = T@c. At the same time, c = Uf where U is ucell with vectors as columns and f is the fractional position. Then c' = T@c = T@Uf = (T@U)f.
--> The right way is just left-multiplying U with T.