fix: remove unwanted side effect
apply_matrix_transform would silently collapse the fractional coordinates after application of the matrix transform. This means that a slab with non-collapsed coordinates would be silently changed to one with collapsed ones.

This seems like an unwanted and undocumented side effect. The collapsing should not be necessary: the transformation should be a rigid one and should not change fractional coordinates.