fix, feat: keep attributes up to date
When a matrix transformation or a unit-cell scaling is applied, make sure we also keep other important attributes up to date:
- bulkslab should be transformed the exact same way to maintain the relationship of coordinates
- layers, sublayers, and bulkslab are surely outdated if the transformation changes the meaning of "z".

The current implementation is consistent with the one in fd_optimization in that the bulkslab does not get its 'origin' updated. It is unclear whether this is actually the right way to proceed.