refactor: replace frozen fortranformat with dependency
We included a frozen version of fortranformat early on because it looked like the project was not maintained anymore. However, this has changed over the last yeast and there have been multiple releases.
This commit removes the frozen version from the repository and adds it as a dependency instead. fortranformat is available from pypi and via conda.