style: fix fixture name conflicts
Fixtures defined in the same module have a problem, mentioned also here https://docs.pytest.org/en/6.2.x/reference.html#fixtures-api

The issue is that the fixture function name is shadowed by the argument names used in the test functions. The better ways are:
- define fixtures in a separate module OR
- use the name= keyword to dynamically rename the fixture.

We opt for the second one for now.