fix: incorrect multiple-fixture parametrization
Using fixtures as parameters is possible, but requires the extra dependency to the request fixture to retrieve the result of executing the fixture body.

For now, dirty solution with explicit use of fixture as an argument. Most likely the best way is a @with_multiple_fixtures decorator that does the mess under the hood.