new: extract common utilities from viperino
Functionality that will be used also for the magnetic-field compensation will be progressively moved into the /lib folder.

This poses an annoying issue when using the Arduino IDE: relative imports are not supported. The solution is to "Sketch/Add file..." in the current "project". This however COPIES the file to the local tree.

The better solution is using the command-line interface. (see next commits)

Co-Authored-By: cpfungen <120585460+cpfungen@users.noreply.github.com>