fix: rfactor interpolation for small steps
This changes the value passed to TensErLEED for the parameter EINCR in the R-factor calculation. Previously, the minimum step width of experimenl and theory would be used, rather than the value defined in IV_SHIFT_RANGE. This meant that for a fine optimization of V0r, the step width used for interpolation would be to coarse. This would result in multiple V0r steps in ROUT giving the same R-factor.

Issue originally pointed out by @LutzHammer.