refactor: parameter interpretation
Rename a bunch of variables, use the stuff from previous commits, minor code improvements here and there, add various notes as TODOs.

Major TODO is to refactor the interpretation into a class so as to split off all the non-standard parameter interpretations into instance methods. the class is ideal to also keep state, and avoid all the mess of "continue" and nested if..elif...else by just using return statements.