fix: no lambda for dynamic methods
The previous solution used some problematic bits:
1) lambda. Should not be stored for any purpose. It's a recipe for disaster. Better stick with well-behaved functools.partial and the like
2) Would recreate the methods for each and every instance during __init__. The methods were unbound.

The correct solution is creating the methods as FUNCTION-type (i.e., descriptor-type) attributes of the CLASS. This way (1) they are automatically bound to whatever instance, (2) they are inheritable (should we ever need to).

The ideal solution for this is (1) create CLASSMETHODS that attach the new methods to the class; (2) functools.partialmethod (instead of .partial) to create the new instance methods as functions on the class.

The disadvantage of this is that we need to use keyword arguments, as partial only allows replacing the early positional arguments. This is a non-issue, as we can still do the exact same thing, provided we re-sort the arguments.

The extra advantage is maintaintability: should we add invalid arguments for new parameters, this will show up at MODULE-IMPORT rather than at runtime. I.e., it's fixed immediately at the first test.