refactor: remove defines wrt registers
Helper function 'pin_to_tc4_reg_addr' is specifically made to perform the conversion from an I/O pin number to the corresponding Timer/Counter4 register address. The user supplies the pin number and the code will return the correct TC4 register address. If the code is migrated to another [Arduino] platform, this function has to be adapted appropriately.