feat, new: utility to write file for VASP
This adds a rudimentary utility (vasp_relax.py) that can take a POSCAR fiel as used by ViPErLEED, i.e. a thick, asymmetric slab and exports it to a format that can be used for a DFT pre-relaxation in VASP. A vacuum gap is added on top of the slab to avoid self interaction and flags are added to only allow atoms above a specified c fraction to relax.