new, tests: helpers module
The file was there before, but was empty. Was then moved to conftest, and fixture have been added.

It makes sense to have fixtures in conftest. However, it seems not so clean to have non-fixture utilities in conftest files.

The downside is that, in order to be able to import from helpers from deeper down the tree, we need __init__.py files at all levels.