tests, new: large test set for symmetry
Add a significantly large number of tests for symmetry detection cases, moved to a package on its own for simplicity.

Many of the tests (~30%) currently FAIL, and are marked as XFAIL where appropriate. All of the issues are actually already fixed in my reimplementation of better_symmetry (to be hopefully soon committed).