fix!: correct FORTRAN format
Few issues in the previous implementation:
- format of beams was swapped: it's I5 in >=1.7 and I4 in earlier versions.
- Format for n_beams has not been I3 since a long time. Was a stray comment in the old beamgen file. The format has ben changed by @LutzHammer twice: I4 in v3, I5 in v1.7. This is now handled corectly.

These are now handled correctly.

Also, importantly, we now raise in case the number of beams exceeds the maximum that the format can handle. This is way better than writing asterisks to file and then have the Fortran code complain with some cryptic message (if any).