refactor: slab package, from_ase, sym_entity
This way we can split functionality into three logical objects: BulkSlab (3D periodic), SurfaceSlab (2D periodic), and BaseSlab (functionality common to both).

While reorganizing, it also makes sense to move SymPlane to its own module, as it is not uniquely used in slabs.

For now, all the methods in slab.py are moved as-they-are in the new classes. A more thorough rewrite will soon come. Only a few minor changes, especially for what concerns method sorting. This is the current order:

- alphabetically ordered public @property
- alphabetically ordered public @classmethod
- alphabetically ordered public instance methods, with the exception of methods that concern geometrical transformation of slabs and symmetry-compliance methods
- alphabetically ordered methods dealing with slab geometric transformations
- alphabetically ordered methods dealing with checks of symmetry properties

Two core differences:
- now makeBulkSlab returns a BulkSlab
- the from_ase interface has changed. It is now a @classmethod of SurfaceSlab (only!)