refactor: make updateElements private
Used only internally in fullUpdate. It is probably good practice not to use it anywhere else.

Also, (i) improve code readability, (ii) replace chemelem with a set (we only use it for 'in' tests or iteration), (iii) **important**: store a copy of ELEMENT_MIX -- it is a mapping that may be modified in-place.

@fkraushofer: could you comment on my note about ELEMENT_RENAME at line 995?