refactor, fix: collapseFractionalCoordinates
Move the implementation to base, as it may be useful in other places.

Also, the implementation is changed slightly:
rather than using % 1.0 for collapsing, allow using two different methods. One rounds to 'closest to zero', the other one is de facto identical to %1.0 but behaves a little better for numbers that are slightly smaller but close to integers (e.g., 1-1e-8; 1-1e-9; ...; 1-1e-15).