refactor, fix: sort_by_z
(1) Use the intrinsic ability of sort to reverse rather then reversing the list of atoms; (2) **important** use the correct meaning for the keyword argument! The earlier implementation would sort the list from bottom to top (lower to higher pos[2]) when botToTop=False. However, we want the sorting to be bottom-to-top if botToTop is **True**.

Changed the default to maintain the same "normal behavior" of sorting bottom-to-top like before.