fix: element-wise comparison
This fix was also introduced by @fkraushofer in #51. The problem with the previous implementation is that checks of translational symmetry would be transparent to the chemical nature of an atom. I.e., moving all Mg atoms onto O atoms would be considered a valid translation symmetry.

However, the fix in #51 has a bug in that element_per_atom_shifted is not restricted to 'in-range' coordinates. This gives a bit of a mess in the index maps.