refactor, feat, fix: addBulkLayers
- rename with_extra_bulk_units. This makes it clearer that the method is actually returning a new object rather than adding stuff to the slab this method is called on
- method moved to SurfaceSlab. Currently, I feel that BulkSlabs probably don't need it (they have doubleBulkSlab). The method can however be moved back as-is into BaseSlab, and would guarantee a finer-grained control on how many cells are added to a bulk slab (as opposed to making twice as thick). I'd be happy to hear @fkraushofer and @amimre opinion about this.
- The method is split into two parts, one that is solely responsible of adding one unit bulk cell (which will soon be used to refactor also doubleBulkSlab).

A few improvements:
- atom numbering is always correct now, avoiding DuplicateAtomsError. The earlier implementation kept track of which atoms were duplicated.
- now calling the method twice, i.e., adding bulk units to a slab to which bulk units had already been added, works correctly. The earlier implementation would fail, partly because it was not adding layers, partly because it assumed that only one or two of the layers would be 'bulk'.

There is one non-ideality, however, that I'd like to discuss with @fkraushofer and @amimre. The method currently maintains all the added layers, as well as those originally present, labelled as 'bulk'. This means that a second call to this method will add 'more bulk' than the first call given the same number of cells. The only sustainable way to fix this is to re-label all layers to non-bulk, always leaving the same number of bulk layers at the bottom. This also makes slabs returned by this method potentially not usable for ViPErLEED refcalc, as we allow at most two bulk layers. On the other hand, turning 'bulk' layers into 'non-bulk' layers also has its disadvantages (e.g., atoms that were not accessible for a fit become accessible. This may or may not be desirable).