feat: don't collapse if nothing to undo
Former implementation of revert_unit_cell (which was the same before the refactor in e31cfa5b66b993f8f82f5895876adff963b93638) would collapse even if the user asked for 'undo nothing'. This seems like a somewhat unnecessary side effect, as 'nothing to undo' feels like it should not modify anything. I can revert this if @fkraushofer and/or @amimre feel like the earlier behavior is desirable.