feat: revert_unit_cell always collapses
After discussion with @fkraushofer, we agreed that the behavior of revert_unit_cell should be of always collapsing the atom coordinates to the base cell.

This commit essentially reverts the edits in 0aec33bfc9db7b34f52e11e4b17c8ed618ccc026. However, it also changes the behavior of the old implementation: There, we would not collapse if the unit cell had no 'registered' modifications. This commit makes the behavior overall consistent.

Co-Authored-By: Florian Kraushofer <66359721+fkraushofer@users.noreply.github.com>